rename app resource to chart
