+++ service/controller/v1/resource/chart/error.go
@@ -1,4 +1,4 @@
-package app
+package chart
 
 import "github.com/giantswarm/microerror"
