+++ service/controller/v1/resource/chart/resource.go
@@ -1,4 +1,4 @@
-package app
+package chart
 
 import (
 	"context"
